requirements.txt
-i https://pypi.org/simple
click==8.1.3; python_version >= '3.7'
flask==2.2.2
flask-login==0.6.2
flask-sqlalchemy==3.0.3
flask-wtf==1.1.1
greenlet==2.0.2; platform_machine == 'aarch64' or (platform_machine == 'ppc64le' or (platform_machine == 'x86_64' or (platform_machine == 'amd64' or (platform_machine == 'AMD64' or (platform_machine == 'win32' or platform_machine == 'WIN32')))))
itsdangerous==2.1.2; python_version >= '3.7'
jinja2==3.1.2; python_version >= '3.7'
markupsafe==2.1.2; python_version >= '3.7'
python-dotenv==0.21.1
sqlalchemy==2.0.1; python_version >= '3.7'
typing-extensions==4.4.0; python_version >= '3.7'
werkzeug==2.2.2; python_version >= '3.7'
wtforms==3.0.1
